Add files via upload
This is the class for login screen we have added a requirement that the user must provide a source file path. We manipulated the gameLauncher class to create tests and test coverage. We had to change the testing class due to lack of knowledge of VCS file control and lack of web development experience.